Split remaining service dependencies into their own modules
This is AWS, Graphite, Jaeger, SignalFx, and Zipkin.

The Zipkin/Jaeger change is unsatisfying because we used to support choosing a tracer
at runtime depending on what was in the config. This was nice behavior but it meant
that every application had the union of all tracing dependencies in their binary.
I'm going to try to find a nicer way in follow-up.
