+++ misk-aws/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-aws
+POM_NAME=misk-aws
+POM_DESCRIPTION=Misk module for AWS
+POM_PACKAGING=jar
