+++ misk-aws/src/test/kotlin/misk/cloud/aws/logging/CloudwatchTrailTest.kt
