+++ misk-graphite/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-graphite
+POM_NAME=misk-graphite
+POM_DESCRIPTION=Misk module for Graphite
+POM_PACKAGING=jar
