+++ misk-jaeger/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-jaeger
+POM_NAME=misk-jaeger
+POM_DESCRIPTION=Misk module for Jaeger
+POM_PACKAGING=jar
