+++ misk-signalfx/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-signalfx
+POM_NAME=misk-signalfx
+POM_DESCRIPTION=Misk module for SignalFx
+POM_PACKAGING=jar
