+++ misk-zipkin/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-zipkin
+POM_NAME=misk-zipkin
+POM_DESCRIPTION=Misk module for Zipkin
+POM_PACKAGING=jar
