--- misk/src/main/kotlin/misk/tracing/TracingConfig.kt
@@ -1,6 +1,0 @@
-package misk.tracing
-
-import misk.config.Config
-import misk.tracing.backends.TracingBackendConfig
-
-data class TracingConfig(val backends: TracingBackendConfig) : Config
