--- misk/src/main/kotlin/misk/tracing/backends/TracingBackendConfig.kt
@@ -1,10 +1,0 @@
-package misk.tracing.backends
-
-import misk.config.Config
-import misk.tracing.backends.jaeger.JaegerBackendConfig
-import misk.tracing.backends.zipkin.ZipkinBackendConfig
-
-data class TracingBackendConfig(
-  val jaeger: JaegerBackendConfig? = null,
-  val zipkin: ZipkinBackendConfig? = null
-) : Config
