--- misk/src/test/kotlin/misk/tracing/TestTracingConfig.kt
@@ -1,7 +1,0 @@
-package misk.tracing
-
-import misk.config.Config
-
-data class TestTracingConfig(
-  val tracing: TracingConfig
-) : Config
