+++ settings.gradle
@@ -1,9 +1,14 @@
 include ':misk'
+include ':misk-aws'
 include ':misk-gcp'
 include ':misk-gcp-testing'
+include ':misk-jaeger'
+include ':misk-graphite'
 include ':misk-hibernate'
 include ':misk-hibernate-testing'
+include ':misk-signalfx'
 include ':misk-testing'
+include ':misk-zipkin'
 include ':samples:exemplar'
 include ':samples:exemplarchat'
 include ':samples:urlshortener'
